update - move components from pages directory to its own directory in components dir
